requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
alembic==1.7.5
    # via -r requirements/base.txt
anyio==3.4.0
    # via
    #   -r requirements/base.txt
    #   starlette
asgiref==3.4.1
    # via
    #   -r requirements/base.txt
    #   uvicorn
asyncpg==0.25.0
    # via -r requirements/base.txt
black==21.11b1
    # via
    #   -r requirements/base.txt
    #   -r requirements/dev.in
certifi==2021.10.8
    # via
    #   -r requirements/base.txt
    #   requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.8
    # via
    #   -r requirements/base.txt
    #   requests
click==8.0.3
    # via
    #   -r requirements/base.txt
    #   black
    #   pip-tools
    #   uvicorn
distlib==0.3.4
    # via virtualenv
dnspython==2.1.0
    # via
    #   -r requirements/base.txt
    #   email-validator
email-validator==1.1.3
    # via
    #   -r requirements/base.txt
    #   fastapi
fastapi[all]==0.70.0
    # via -r requirements/base.txt
filelock==3.4.2
    # via virtualenv
flake8==4.0.1
    # via -r requirements/dev.in
greenlet==1.1.2
    # via
    #   -r requirements/base.txt
    #   sqlalchemy
h11==0.12.0
    # via
    #   -r requirements/base.txt
    #   uvicorn
httptools==0.2.0
    # via
    #   -r requirements/base.txt
    #   uvicorn
identify==2.4.5
    # via pre-commit
idna==3.3
    # via
    #   -r requirements/base.txt
    #   anyio
    #   email-validator
    #   requests
importlib-metadata==4.8.2
    # via
    #   -r requirements/base.txt
    #   alembic
importlib-resources==5.4.0
    # via
    #   -r requirements/base.txt
    #   alembic
isort==5.10.1
    # via -r requirements/dev.in
itsdangerous==2.0.1
    # via
    #   -r requirements/base.txt
    #   fastapi
jinja2==3.0.3
    # via
    #   -r requirements/base.txt
    #   fastapi
mako==1.1.6
    # via
    #   -r requirements/base.txt
    #   alembic
markupsafe==2.0.1
    # via
    #   -r requirements/base.txt
    #   jinja2
    #   mako
mccabe==0.6.1
    # via flake8
mypy==0.931
    # via -r requirements/dev.in
mypy-extensions==0.4.3
    # via
    #   -r requirements/base.txt
    #   black
    #   mypy
nodeenv==1.6.0
    # via pre-commit
numpy==1.21.4
    # via -r requirements/base.txt
orjson==3.6.4
    # via
    #   -r requirements/base.txt
    #   fastapi
pathspec==0.9.0
    # via
    #   -r requirements/base.txt
    #   black
pep517==0.12.0
    # via pip-tools
pillow==8.4.0
    # via -r requirements/base.txt
pip-tools==6.4.0
    # via -r requirements/dev.in
platformdirs==2.4.0
    # via
    #   -r requirements/base.txt
    #   black
    #   virtualenv
pre-commit==2.17.0
    # via -r requirements/dev.in
pycodestyle==2.8.0
    # via flake8
pydantic==1.8.2
    # via
    #   -r requirements/base.txt
    #   fastapi
    #   sqlmodel
pyflakes==2.4.0
    # via flake8
python-dotenv==0.19.2
    # via
    #   -r requirements/base.txt
    #   uvicorn
python-multipart==0.0.5
    # via
    #   -r requirements/base.txt
    #   fastapi
pyyaml==5.4.1
    # via
    #   -r requirements/base.txt
    #   fastapi
    #   pre-commit
    #   uvicorn
regex==2021.11.10
    # via
    #   -r requirements/base.txt
    #   black
requests==2.26.0
    # via
    #   -r requirements/base.txt
    #   fastapi
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   python-multipart
    #   virtualenv
sniffio==1.2.0
    # via
    #   -r requirements/base.txt
    #   anyio
sqlalchemy==1.4.27
    # via
    #   -r requirements/base.txt
    #   alembic
    #   sqlmodel
sqlalchemy2-stubs==0.0.2a19
    # via
    #   -r requirements/base.txt
    #   sqlmodel
sqlmodel==0.0.4
    # via -r requirements/base.txt
starlette==0.16.0
    # via
    #   -r requirements/base.txt
    #   fastapi
toml==0.10.2
    # via pre-commit
tomli==1.2.2
    # via
    #   -r requirements/base.txt
    #   black
    #   mypy
    #   pep517
typing-extensions==4.0.0
    # via
    #   -r requirements/base.txt
    #   black
    #   mypy
    #   pydantic
    #   sqlalchemy2-stubs
ujson==4.3.0
    # via
    #   -r requirements/base.txt
    #   fastapi
urllib3==1.26.7
    # via
    #   -r requirements/base.txt
    #   requests
uvicorn[standard]==0.15.0
    # via
    #   -r requirements/base.txt
    #   fastapi
uvloop==0.16.0
    # via
    #   -r requirements/base.txt
    #   uvicorn
virtualenv==20.13.0
    # via pre-commit
watchgod==0.7
    # via
    #   -r requirements/base.txt
    #   uvicorn
websockets==10.1
    # via
    #   -r requirements/base.txt
    #   uvicorn
wheel==0.37.1
    # via pip-tools
zipp==3.6.0
    # via
    #   -r requirements/base.txt
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
